requirements.txt
asgiref==3.9.1
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
bs4==0.0.2
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
coverage==7.10.2
Django==5.2.4
django-cors-headers==4.7.0
django-debug-toolbar==6.0.0
django-filter==25.1
djangorestframework==3.16.1
drf-yasg==1.21.10
et_xmlfile==2.0.0
Flask==3.1.1
# Editable install with no version control (flaskr==1.0.0)
-e c:\users\asus\desktop\°³¹ß\flask\projects\flask-tutorial
h11==0.16.0
idna==3.10
image==1.5.33
inflection==0.5.1
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
keyboard==0.13.5
MarkupSafe==3.0.2
MouseInfo==0.1.3
mypy==1.17.1
mypy_extensions==1.1.0
mysqlclient==2.2.7
numpy==2.3.1
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.1
pathspec==0.12.1
pillow==11.3.0
pluggy==1.6.0
PyAutoGUI==0.9.54
pycparser==2.22
PyGetWindow==0.0.9
Pygments==2.19.2
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytweening==1.2.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
selenium==4.34.2
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
sqlparse==0.5.3
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
urlopen==1.0.0
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
